# docs/swagger.yaml
openapi: "3.0.0"
servers:
  - url: 'https://ubademy-14-prod.herokuapp.com/'
info:
  description: Course API encargada de todas las operaciones relacionadas a cursos
  version: 0.2.0
  title: Course

tags:
  - name: Course
    description: Operaciones sobre cursos

paths:
  '/courses/create':
    'post':
      tags:
        - Course
      description: Crea un curso nuevo
      parameters:
        - in: header
          name: x-access-token
          required: true
          description: token recibido del usuario que desea crear el curso
      requestBody:
        description: body para la creacion del curso
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/courseCreate'
      responses:
        '200':
          description: Curso creado correctamente
        '400':
          description: Bad Request
        '401':
          description: Usuario bloqueado

  '/courses':
    'get':
      tags:
        - Course
      description: Obtengo una lista con todos los cursos (que NO estan CANCELADOS o BLOQUEADOS) que su estado es on course
      parameters:
        - in: query
          name: location
          example: 'Obera'
        - in: query
          name: name
          example: 'Python'
        - in: query
          name: creator_first_name
          example: 'Juan'
        - in: query
          name: creator_last_name
          example: 'Perez'
        - in: query
          name: hashtags
          example: 'python, nacho, branca'
        - in: query
          name: type
          example: 'Programacion'
        - in: query
          name: subscription
          example: 'Premium'
        - in: query
          name: exams
          example: 5
        - in: query
          name: free_text
          example: 'tres tristes tigres'
        - in: query
          name: offset
          example: 0
        - in: query
          name: limit
          example: 500
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario

      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/coursesList'
        '204':
          description: No content
        '400':
          description: Bad Request

  '/courses/{id}/view':
    'get':
      tags:
        - Course
      description: Obtiene la informacion de un curso. Si el curso esta CANCELADO solo el CREADOR lo puede ver. Si esta BLOQUEADO solo el ADMIN puede verlo.
      parameters:
        - in: path
          name: id
          description: id del curso
          required: true
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/courseInfo'
        '404':
          description: Curso no existe

  '/courses/{id}':
    'delete':
      tags:
        - Course
      description: Cancela un curso existente
      parameters:
        - in: path
          name: id
          description: id del curso
          required: true
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Curso eliminado correctamente
        '400':
          description: Invalid user action
        '401':
          description: Token invalido
        '403':
          description: No token provisto
        '404':
          description: Curso no encontrado

    'patch':
      tags:
        - Course
      description: Edita la informacion de un curso
      parameters:
        - in: path
          name: id
          description: id del curso
          example: 1
          required: true
        - in: header
          name: x-access-token
          required: true
          description: token recibido del usuario que desea editar el curso
      requestBody:
        description: Body para la edicion del curso
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/editCourse'
      responses:
        '200':
          description: Curso editado correctamente
        '400':
          description: Invalid info
        '401':
          description: Invalid user action
        '403':
          description: No token provisto
        '404':
          description: Course does not exist
        '405':
          description: El usuario no es el creador del curso

  '/courses/collaborators':
    'post':
      tags:
        - Course
      description: Agrega un colaborador al curso
      parameters:
        - in: header
          name: x-access-token
          required: true
          description: token recibido del usuario que sera colaborador
      requestBody:
        description: Body para agregar un colaborador
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/courseAddCollaborator'
      responses:
        '200':
          description: Collaborator agregado correctamente
        '208':
          description: Ya es un colaborador del curso
        '401':
          description: Token invalido
        '403':
          description: No token provisto
        '404':
          description: El curso no existe

  '/courses/collaborators/send_request':
    'post':
      tags:
        - Course
      description: Envia la solicitud para ser colaborador de un curso
      parameters:
        - in: header
          name: x-access-token
          required: true
          description: token recibido del creador del curso
      requestBody:
        description: Body para enviar la solicitud
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sendCollaborationRequest'
      responses:
        '200':
          description: Solicitud enviada correctamente
        '208':
          description: Ya es un colaborador del curso
        '401':
          description: Token invalido
        '403':
          description: No token provisto
        '404':
          description: El curso no existe o el usuario no existe

  '/courses/collaborators/remove':
    'delete':
      tags:
        - Course
      description: Elimina a un colaborador del curso
      parameters:
        - in: header
          name: x-access-token
          required: true
          description: token recibido del usuario que realiza la accion
      requestBody:
        description: Body para quitar un colaborador
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/courseRemoveCollaborator'
      responses:
        '200':
          description: Collaborator removido correctamente
        '401':
          description: Token invalido
        '403':
          description: No token provisto
        '404':
          description: El collaborator no existe

  '/courses/subscription/{id}':
    'post':
      tags:
        - Course
      description: Agrega un suscriptor al curso
      parameters:
        - in: path
          name: id
          description: id del curso
        - in: header
          name: x-access-token
          required: true
          description: token recibido del usuario que quiere suscribirse
      responses:
        '200':
          description: Suscriptor agregado correctamente
        '208':
          description: El usuario ya esta suscrito
        '401':
          description: Token invalido
        '403':
          description: No token provisto o You should upgrade your membership to enroll this course
        '404':
          description: El curso no existe

    'delete':
      tags:
        - Course
      description: Elimina a un suscriptor del curso
      parameters:
        - in: path
          name: id
          description: id del curso
        - in: header
          name: x-access-token
          required: true
          description: token recibido del usuario que realiza la accion

      responses:
        '200':
          description: Suscriptor eliminado correctamente
        '401':
          description: Token invalido
        '403':
          description: No token provisto
        '404':
          description: El curso no existe

  '/courses/my_courses':
    'get':
      tags:
        - Course
      description: Obtiene los cursos (que NO estan BLOQUEADOS) que cree
      parameters:
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/myCourses'
        '204':
          description: No content
        '401':
          description: Usuario bloqueado

  '/courses/my_subscriptions':
    'get':
      tags:
        - Course
      description: Obtiene los cursos (que NO estan CANCELADOS o BLOQUEADOS) que estoy cursando acutalmente (status En Curso)
      parameters:
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mySubscriptions'
        '204':
          description: No content

  '/notification':
    'post':
      tags:
        - Notification
      description: Envia una notificacion a un usuario
      requestBody:
        description: Body para enviar la notificacion
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/notification'
      responses:
        '200':
          description: Notificacion enviada correctamente
        '400':
          description: Bad request
        '404':
          description: El usuario no existe

  '/courses/users/{id}':
    'get':
      tags:
        - Course
      description: Obtiene los alumnos o colaboradores (que NO estan BLOQUEADOS) de un curso
      parameters:
        - in: path
          name: id
          description: id del curso
        - in: query
          name: first_name
          example: 'Juan'
        - in: query
          name: last_name
          example: 'Perez'
        - in: query
          name: subscribers
          description: Si es true obtiene los subscribers, caso contrario devuelve los colaboradores de un curso
          example: 'true'
        - in: query
          name: offset
          example: 0
        - in: query
          name: limit
          example: 100
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/courseUsers'
        '204':
          description: No content

  '/courses/block/{id}':
    'delete':
      tags:
        - Course
      description: Bloquea un curso
      parameters:
        - in: path
          name: id
          description: id del curso
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Curso bloqueado correctamente
        '208':
          description: El curso ya se encuentra bloqueado
        '400':
          description: Invalid user action
        '404':
          description: Course not found

  '/courses/unblock/{id}':
    'post':
      tags:
        - Course
      description: Desbloquea un curso
      parameters:
        - in: path
          name: id
          description: id del curso
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Curso desbloqueado correctamente
        '208':
          description: El curso ya se encuentra desbloqueado
        '400':
          description: Invalid user action
        '404':
          description: Course not found

  '/courses/favorites':
    'post':
      tags:
        - Course
      description: Agrega un curso favorito al usuario
      parameters:
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      requestBody:
        description: Body para agregar un curso favorito
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/favoriteCourse'
      responses:
        '200':
          description: Curso faveado correctamente
        '208':
          description: El curso ya se encuentra faveado
        '404':
          description: Course not found

    'get':
      tags:
        - Course
      description: Obtiene los cursos favoritos (que NO estan CANCELADOS o BLOQUEADOS)
      parameters:
        - in: query
          name: location
          example: 'Obera'
        - in: query
          name: name
          example: 'Python'
        - in: query
          name: creator_first_name
          example: 'Juan'
        - in: query
          name: creator_last_name
          example: 'Perez'
        - in: query
          name: hashtags
          example: 'python, nacho, branca'
        - in: query
          name: type
          example: 'Programacion'
        - in: query
          name: subscription
          example: 'Premium'
        - in: query
          name: exams
          example: 5
        - in: query
          name: free_text
          example: 'tres tristes tigres'
        - in: query
          name: offset
          example: 0
        - in: query
          name: limit
          example: 500
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/favCourses'
        '400':
          description: Bad Request

  '/courses/favorites/remove':
    'delete':
      tags:
        - Course
      description: Quita un curso de la lista de favoritos
      parameters:
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      requestBody:
        description: Body para quitar un curso faveado
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/favoriteCourse'
      responses:
        '200':
          description: Curso eliminado correctamente de los favoritos
        '208':
          description: El curso no se encuentra faveado
        '400':
          description: Invalid user action
        '404':
          description: Course not found

  '/courses/historical':
    'get':
      tags:
        - Course
      description: Obtiene el historico de cursos (que NO estan CANCELADOS o BLOQUEADOS). En esta lista figuran los cursos con estado Aprobado, Desaprobado o En Curso
      parameters:
        - in: query
          name: status
          example: 'Aprobado'
        - in: query
          name: offset
          example: 0
        - in: query
          name: limit
          example: 500
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/historical'
        '204':
          description: No content
        '400':
          description: Bad Request

  '/courses/my_collaborations':
    'get':
      tags:
        - Course
      description: Obtiene todos los cursos (que NO estan CANCELADOS o BLOQUEADOS) en los cuales soy colaborador
      parameters:
        - in: query
          name: location
          example: 'Obera'
        - in: query
          name: name
          example: 'Python'
        - in: query
          name: creator_first_name
          example: 'Juan'
        - in: query
          name: creator_last_name
          example: 'Perez'
        - in: query
          name: hashtags
          example: 'python, nacho, branca'
        - in: query
          name: type
          example: 'Programacion'
        - in: query
          name: subscription
          example: 'Premium'
        - in: query
          name: exams
          example: 5
        - in: query
          name: free_text
          example: 'tres tristes tigres'
        - in: query
          name: offset
          example: 0
        - in: query
          name: limit
          example: 500
        - in: header
          name: x-access-token
          required: true
          description: token recibido de usuario
      responses:
        '200':
          description: Todo okay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/collaborations'
        '204':
          description: No content
        '400':
          description: Bad Request

components:
  schemas:
    courseCreate:
      type: object
      required:
        - name
        - description
        - hashtags
        - type
        - amount_exams
        - subscription
        - location
      properties:
        name:
          type: string
          example: 'Python'
        description:
          type: string
          example: 'Van a codear muchas cosas en python y van a ver que es alto lenguaje'
        hashtags:
          type: string
          example: 'Python, Sarasa'
        type:
          type: string
          example: 'Programacion'
        exams:
          type: integer
          example: 3
        subscription:
          type: string
          example: 'Premium'
        location:
          type: string
          example: 'Obera'

    coursesList:
      type: object
      properties:
        name:
          type: string
          example: 'Python'
        id:
          type: integer
          example: 1
        creator_id:
          type: integer
          example: 10
        creator_first_name:
          type: string
          example: 'Juan'
        creator_last_name:
          type: string
          example: 'Perez'
        can_edit:
          type: boolean
          example: true
        can_collaborate:
          type: boolean
          example: true
        can_subscribe:
          type: boolean
          example: false
        subscription:
          type: string
          example: 'premium'
        location:
          type: string
          example: 'Obera'
        hashtags:
          type: string
          example: 'Python, Sarasa'
        desciption:
          type: string
          example: 'Van a codear muchas cosas en python y van a ver que es alto lenguaje'
        exams:
          type: integer
          example: 1
        type:
          type: string
          example: 'Programacion'
        cancelled:
          type: integer
          example: 0
        blocked:
          type: boolean
          example: false
        created_at:
          type: string
          example: 'timestamp, swagger no tiene'
        updated_at:
          type: string
          example: 'timestamp, swagger no tiene'

    editCourse:
      type: object
      properties:
        name:
          type: string
          example: 'El regreso de Python'
        description:
          type: string
          example: 'Van a codear muchas cosas en python y van a ver que es alto lenguaje'
        hashtags:
          type: string
          example: 'Python, Sarasa, Fulano'
        location:
          type: string
          example: 'Corrientes'

    courseInfo:
      type: object
      properties:
        name:
          type: string
          example: 'Python'
        id:
          type: integer
          example: 1
        creator_id:
          type: integer
          example: 10
        creator_first_name:
          type: string
          example: 'Juan'
        creator_last_name:
          type: string
          example: 'Perez'
        can_edit:
          type: boolean
          example: true
        can_collaborate:
          type: boolean
          example: true
        can_subscribe:
          type: boolean
          example: false
        is_subscribed:
          type: boolean
          example: false
        subscription:
          type: string
          example: 'premium'
        location:
          type: string
          example: 'Obera'
        hashtags:
          type: string
          example: 'Python, Sarasa'
        desciption:
          type: string
          example: 'Van a codear muchas cosas en python y van a ver que es alto lenguaje'
        exams:
          type: integer
          example: 1
        type:
          type: string
          example: 'Programacion'
        cancelled:
          type: integer
          example: 0
        blocked:
          type: boolean
          example: false
        created_at:
          type: string
          example: 'timestamp, swagger no tiene'
        updated_at:
          type: string
          example: 'timestamp, swagger no tiene'

    courseAddCollaborator:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          description: id del curso
          example: 1

    courseRemoveCollaborator:
      type: object
      required:
        - id_creator
        - id
        - collaborator_mail
      properties:
        id:
          type: integer
          description: id del curso
          example: 1
        user_to_remove:
          type: integer
          description: id del colaborador a eliminar
          example: 50

    myCourses:
      type: object
      properties:
        myCourses:
          type: array
          description: Es un array en el cual cada elemento contiene TODA la informacion del curso que cree

    mySubscriptions:
      type: object
      properties:
        mySubscriptions:
          type: array
          description: Es un array en el cual cada elemento contiene TODA la informacion del curso al cual me suscribi

    courseUsers:
      type: object
      properties:
        courseUsers:
          type: array
          description: Es un array en el cual cada elemento contiene TODA la informacion de un usuario

    favoriteCourse:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          description: id del curso a favear
          example: 1

    favCourses:
      type: object
      properties:
        favCourses:
          type: array
          description: Es un array en el cual cada elemento contiene TODA la informacion de un curso faveado

    historical:
      type: object
      properties:
        historical:
          type: array
          description: Es un array en el cual cada elemento contiene TODA la informacion de un curso que aprobe, desaprobe o estoy realizando

    collaborations:
      type: object
      properties:
        collaborations:
          type: array
          description: Es un array en el cual cada elemento contiene TODA la informacion de un curso en el cual soy colaborador

    sendCollaborationRequest:
      type: object
      required:
        - id
        - user_id
        - email
      properties:
        id:
          type: integer
          description: Id del curso
          example: 1
        user_id:
          type: integer
          description: Id del usuario que decide enviar la solicitud
          example: 3
        email:
          type: string
          description: email del usuario al que se le va a enviar la solicitud
          example: 'ubademy.14@gmail.com'

    notification:
      type: object
      required:
        - title
        - body
        - user_id
      properties:
        title:
          type: string
          description: Titulo de la notificacion
          example: 'Temaiken'
        body:
          type: string
          description: Id del usuario que decide enviar la solicitud
          example: 'Hola que tal tu como estas? dime si eres feliz'
        user_id:
          type: integer
          description: id del usuario al que se le envia la notificacion
          example: 3